UPDATE: - Move logic to AccountController
