+++ client/src/services/UserService.js
@@ -15,18 +15,18 @@ export class UserService {
 	 * Move into SW
 	 */
 	async generateKeypair(loginInfo = {}) {
-		try {
-			const currentWalletIndex = loginInfo?.currentWalletIndex ?? 0;
-			const encryptionType = loginInfo?.encryptionType ?? 'Ed25519';
-			const user = this.instance;
+		// try {
+		// 	const currentWalletIndex = loginInfo?.currentWalletIndex ?? 0;
+		// 	const encryptionType = loginInfo?.encryptionType ?? 'Ed25519';
+		// 	const user = this.instance;
 
-			const wallet = await user.getWalletAccount(currentWalletIndex);
-			const publicKey = await wallet.getPublicKeyByteArray();
-			const secretKey = wallet.getPrivateKeyByteArray();
-			return Keys[encryptionType].parseKeyPair(publicKey.slice(1), secretKey);
-		} catch (error) {
-			return undefined;
-		}
+		// 	const wallet = await user.getWalletAccount(currentWalletIndex);
+		// 	const publicKey = await wallet.getPublicKeyByteArray();
+		// 	const secretKey = wallet.getPrivateKeyByteArray();
+		// 	return Keys[encryptionType].parseKeyPair(publicKey.slice(1), secretKey);
+		// } catch (error) {
+		// 	return undefined;
+		// }
 	}
 }
